FY-131/refacor: useFetchProfileData 커스텀 훅 작성 및 적용
